Don not delete extension keys while probe is still connected, code refactoring.

+++ biomio/protocol/rpc/app_auth_connection.py
@@ -53,22 +53,17 @@ class AppAuthConnection():
         if len(connected_apps):
             AppConnectionManager.instance().add_active_app(connected_apps[0], self._app_id)
             if self.is_probe_owner():
-                self._app_key = self._listener.auth_key(extension_id=connected_apps[0], probe_id=self._app_id)
-                app_key_pattern = AppConnectionListener.app_key_pattern(app_id=self._app_id, app_type=self._app_type)
-                existing_keys = AuthStateStorage.instance().get_matching_keys(pattern=app_key_pattern)
-                if self._app_key in existing_keys:
-                    existing_keys.remove(self._app_key)
-                AuthStateStorage.instance().remove_keys(keys=existing_keys)
-                extension_id = AppConnectionListener.extension_id(redis_auth_key=self._app_key)
-                pattern = AppConnectionListener.app_key_pattern(app_id=extension_id, app_type='extension')
-                keys_to_remove = AuthStateStorage.instance().get_matching_keys(pattern)
-                if self._app_key in keys_to_remove:
-                    keys_to_remove.remove(self._app_key)
-                AuthStateStorage.instance().remove_keys(keys=keys_to_remove)
+                new_app_key = self._listener.auth_key(extension_id=connected_apps[0], probe_id=self._app_id)
             else:
-                self._app_key = self._listener.auth_key(extension_id=self._app_id, probe_id=connected_apps[0])
-            self.extension_keys = [self._app_key]
-            self.store_data(state='auth_wait')
+                new_app_key = self._listener.auth_key(extension_id=self._app_id, probe_id=connected_apps[0])
+            if self._app_key is None or self._app_key != new_app_key:
+                self.remove_extension_keys_that_are_not_connected()
+                self._app_key = new_app_key
+                self.extension_keys = []
+            if on_behalf_of is not None:
+                if not len(self.extension_keys):
+                    self.extension_keys = [self._app_key]
+                self.store_data(state='auth_wait')
         else:
             app_id = self._app_id if on_behalf_of is None else on_behalf_of
             AppConnectionManager.get_connected_apps(app_id=app_id, device=self.is_probe_owner(),
@@ -110,22 +105,21 @@ class AppAuthConnection():
         self._listener.subscribe(callback=self._on_connection_data)
         if self.is_probe_owner():
             self._set_keys_for_connected_app()
-        # if self.is_probe_owner():
-        #     app_key_pattern = AppConnectionListener.app_key_pattern(app_id=self._app_id, app_type=self._app_type)
-        #     existing_keys = AuthStateStorage.instance().get_matching_keys(pattern=app_key_pattern)
-        #     if existing_keys:
-        #         self._app_key = existing_keys.pop(0)
-        #     AuthStateStorage.instance().remove_keys(keys=existing_keys)
-        #     self.remove_extension_keys_that_are_not_connected()
+            app_key_pattern = AppConnectionListener.app_key_pattern(app_id=self._app_id, app_type=self._app_type)
+            existing_keys = AuthStateStorage.instance().get_matching_keys(pattern=app_key_pattern)
+            if self._app_key in existing_keys:
+                existing_keys.remove(self._app_key)
+            AuthStateStorage.instance().remove_keys(keys=existing_keys)
+            self.remove_extension_keys_that_are_not_connected()
 
-    # def remove_extension_keys_that_are_not_connected(self):
-    #     if self._app_key:
-    #         extension_id = AppConnectionListener.extension_id(redis_auth_key=self._app_key)
-    #         pattern = AppConnectionListener.app_key_pattern(app_id=extension_id, app_type='extension')
-    #         keys_to_remove = AuthStateStorage.instance().get_matching_keys(pattern)
-    #         if self._app_key in keys_to_remove:
-    #             keys_to_remove.remove(self._app_key)
-    #         AuthStateStorage.instance().remove_keys(keys=keys_to_remove)
+    def remove_extension_keys_that_are_not_connected(self):
+        if self._app_key:
+            extension_id = AppConnectionListener.extension_id(redis_auth_key=self._app_key)
+            pattern = AppConnectionListener.app_key_pattern(app_id=extension_id, app_type='extension')
+            keys_to_remove = AuthStateStorage.instance().get_matching_keys(pattern)
+            if self._app_key in keys_to_remove:
+                keys_to_remove.remove(self._app_key)
+            AuthStateStorage.instance().remove_keys(keys=keys_to_remove)
 
     def set_app_disconnected(self, on_behalf_of=None):
         """
@@ -180,25 +174,23 @@ class AppAuthConnection():
             probe_id=self._app_id if self.is_probe_owner() else other_id
         )
 
-    def _remove_disconnected_app_keys(self, on_behalf_of=None):
+    def _remove_disconnected_app_keys(self, on_behalf_of=None, clear_related_keys=True):
         app_id = self._app_id if on_behalf_of is None else on_behalf_of
+        self.clear_related_keys = clear_related_keys
         AppConnectionManager.get_connected_apps(app_id=app_id, device=self.is_probe_owner(),
                                                 callback=self._get_disconnect_apps_callback())
 
     def _get_disconnect_apps_callback(self):
         def _disconnect_apps(app_ids):
-            app_ids = app_ids.get('result') if app_ids is not None else []
-            logger.debug('App ids to remove connections for: %s' % app_ids)
-            for app_id in app_ids:
-                AppConnectionManager.instance().remove_active_app(app_id, self._app_id)
-            AppConnectionManager.instance().delete_active_apps_list(self._app_id)
+            if self.clear_related_keys:
+                app_ids = app_ids.get('result') if app_ids is not None else []
+                logger.debug('App ids to remove connections for: %s' % app_ids)
+                for app_id in app_ids:
+                    AppConnectionManager.instance().remove_active_app(app_id, self._app_id)
+            if self.is_probe_owner():
+                AppConnectionManager.instance().delete_active_apps_list(self._app_id)
             if self.is_probe_owner() and self._app_key:
-                extension_id = AppConnectionListener.extension_id(redis_auth_key=self._app_key)
-                pattern = AppConnectionListener.app_key_pattern(app_id=extension_id, app_type='extension')
-                keys_to_remove = AuthStateStorage.instance().get_matching_keys(pattern)
-                if self._app_key in keys_to_remove:
-                    keys_to_remove.remove(self._app_key)
-                AuthStateStorage.instance().remove_keys(keys=keys_to_remove)
+                self.remove_extension_keys_that_are_not_connected()
         return _disconnect_apps
 
     def _on_connection_data(self, connected_extension_id, connected_probe_id):
@@ -212,7 +204,7 @@ class AppAuthConnection():
             if AuthStateStorage.instance().probe_data_exists(id=key):
                 self._app_key = key
                 if self.is_probe_owner():
-                    self._remove_disconnected_app_keys()
+                    self._remove_disconnected_app_keys(clear_related_keys=False)
                 # else:
                 #     self.extension_keys = []
